Move schema validation to rule
